refactor: extract 3 largest modules from index.js (-1,855 lines)
Extract createComplianceReporter (688 lines), createABExperimentRunner (745 lines),
and createTrustScoreEngine (668 lines) from the 12,677-line monolithic index.js
into separate files, following the same pattern as honeypot-injector.js and
session-risk-aggregator.js.

Before: index.js = 12,677 lines (single file with 35+ factory functions)
After:  index.js = 10,822 lines + 3 focused modules

Each extracted module:
- Is fully self-contained (shared helpers inlined at top)
- Exports via module.exports for require() compatibility
- Imported back into index.js via require()
- All 177 module-specific tests pass with zero regressions

No behavioral changes — all existing tests pass (22 pre-existing
failures in unrelated JSDOM/batch tests remain unchanged).
