+++ src/index.js
@@ -8158,668 +8158,6 @@ function createLoadTester(options) {
 
 // ── A/B Experiment Runner ────────────────────────────────────────────
 
-/**
- * Factory: create an A/B experiment runner for CAPTCHA configurations.
- *
- * Supports full experiment lifecycle: create experiments with named variants,
- * deterministically assign users via hashing, track solve/fail/abandon events,
- * compute statistical significance (chi-squared + z-test for proportions),
- * detect winners, and support early stopping.
- *
- * @param {Object} [options]
- * @param {number} [options.maxExperiments=50] - Maximum concurrent experiments
- * @param {number} [options.significanceLevel=0.05] - P-value threshold for significance
- * @param {number} [options.minSampleSize=30] - Minimum observations per variant before analysis
- * @param {boolean} [options.earlyStoppingEnabled=true] - Allow early stopping when significance reached
- * @param {number} [options.earlyStoppingConfidence=0.01] - Stricter p-value for early stop
- * @returns {Object} A/B experiment runner API
- *
- * @example
- *   var runner = gifCaptcha.createABExperimentRunner();
- *   runner.createExperiment('diff-test', {
- *     control: { difficulty: 'easy', theme: 'default' },
- *     variants: [
- *       { name: 'hard-dark', config: { difficulty: 'hard', theme: 'dark' } },
- *     ],
- *     targetSampleSize: 200,
- *   });
- *   var variant = runner.assignUser('diff-test', 'user-123');
- *   runner.recordEvent('diff-test', 'user-123', 'solve', { timeMs: 3200 });
- *   var results = runner.analyzeExperiment('diff-test');
- */
-function createABExperimentRunner(options) {
-  options = options || {};
-  var maxExperiments = options.maxExperiments > 0 ? options.maxExperiments : 50;
-  var defaultSignificance = typeof options.significanceLevel === 'number' ? options.significanceLevel : 0.05;
-  var defaultMinSample = options.minSampleSize > 0 ? options.minSampleSize : 30;
-  var earlyStoppingEnabled = options.earlyStoppingEnabled !== false;
-  var earlyStoppingConfidence = typeof options.earlyStoppingConfidence === 'number' ? options.earlyStoppingConfidence : 0.01;
-
-  var experiments = {};
-  var experimentOrder = new LruTracker();
-  var onResultCallbacks = [];
-
-  // ── Helpers ──
-
-  /** djb2 hash for deterministic variant assignment */
-  function djb2(str) {
-    var hash = 5381;
-    for (var i = 0; i < str.length; i++) {
-      hash = ((hash << 5) + hash + str.charCodeAt(i)) & 0x7FFFFFFF;
-    }
-    return hash;
-  }
-
-  /** Approximation of standard normal CDF (Abramowitz & Stegun 26.2.17) */
-  function normalCdf(x) {
-    if (x < -8) return 0;
-    if (x > 8) return 1;
-    var sign = x < 0 ? -1 : 1;
-    x = Math.abs(x);
-    var t = 1 / (1 + 0.2316419 * x);
-    var d = 0.3989422804014327; // 1/sqrt(2*pi)
-    var prob = d * Math.exp(-0.5 * x * x) *
-      (t * (0.319381530 + t * (-0.356563782 + t * (1.781477937 + t * (-1.821255978 + t * 1.330274429)))));
-    return sign < 0 ? prob : 1 - prob;
-  }
-
-  /** Two-tailed z-test p-value for difference in proportions */
-  function proportionZTest(n1, s1, n2, s2) {
-    if (n1 === 0 || n2 === 0) return { z: 0, pValue: 1 };
-    var p1 = s1 / n1;
-    var p2 = s2 / n2;
-    var pPool = (s1 + s2) / (n1 + n2);
-    if (pPool === 0 || pPool === 1) return { z: 0, pValue: 1 };
-    var se = Math.sqrt(pPool * (1 - pPool) * (1 / n1 + 1 / n2));
-    if (se === 0) return { z: 0, pValue: 1 };
-    var z = (p1 - p2) / se;
-    var pValue = 2 * (1 - normalCdf(Math.abs(z)));
-    return { z: z, pValue: pValue };
-  }
-
-  /** Chi-squared test for k variants (Pearson's goodness-of-fit) */
-  function chiSquaredTest(variants) {
-    var totalSolves = 0;
-    var totalAttempts = 0;
-    for (var i = 0; i < variants.length; i++) {
-      totalSolves += variants[i].solves;
-      totalAttempts += variants[i].attempts;
-    }
-    if (totalAttempts === 0) return { chiSq: 0, df: 0, pValue: 1 };
-    var expectedRate = totalSolves / totalAttempts;
-    var chiSq = 0;
-    for (var j = 0; j < variants.length; j++) {
-      var n = variants[j].attempts;
-      if (n === 0) continue;
-      var expectedSolves = n * expectedRate;
-      var expectedFails = n * (1 - expectedRate);
-      if (expectedSolves > 0) {
-        var dS = variants[j].solves - expectedSolves;
-        chiSq += (dS * dS) / expectedSolves;
-      }
-      var fails = n - variants[j].solves;
-      if (expectedFails > 0) {
-        var dF = fails - expectedFails;
-        chiSq += (dF * dF) / expectedFails;
-      }
-    }
-    var df = variants.length - 1;
-    // Approximate chi-squared p-value using Wilson-Hilferty
-    var pValue = 1;
-    if (df > 0 && chiSq > 0) {
-      var k = df;
-      var z2 = Math.pow(chiSq / k, 1 / 3) - (1 - 2 / (9 * k));
-      z2 = z2 / Math.sqrt(2 / (9 * k));
-      pValue = 1 - normalCdf(z2);
-    }
-    return { chiSq: chiSq, df: df, pValue: pValue };
-  }
-
-  /** Compute per-variant stats */
-  function computeVariantStats(variantData) {
-    var completedAttempts = variantData.solves + variantData.fails;
-    var totalAttempts = completedAttempts + variantData.abandons;
-    var solveRate = completedAttempts > 0 ? variantData.solves / completedAttempts : 0;
-    var times = variantData.solveTimes;
-    var avgTime = _mean(times);
-    var medianTime = _median(times);
-    var p95Time = 0;
-    if (times.length > 0) {
-      var sorted = times.slice().sort(_numAsc);
-      p95Time = sorted[Math.floor(sorted.length * 0.95)];
-    }
-    return {
-      name: variantData.name,
-      config: variantData.config,
-      attempts: completedAttempts,
-      totalAttempts: totalAttempts,
-      solves: variantData.solves,
-      fails: variantData.fails,
-      abandons: variantData.abandons,
-      solveRate: solveRate,
-      avgSolveTimeMs: Math.round(avgTime),
-      medianSolveTimeMs: Math.round(medianTime),
-      p95SolveTimeMs: Math.round(p95Time),
-      abandonRate: totalAttempts > 0 ? variantData.abandons / totalAttempts : 0,
-    };
-  }
-
-  // ── Core API ──
-
-  /**
-   * Create a new experiment.
-   * @param {string} experimentId - Unique experiment identifier
-   * @param {Object} spec
-   * @param {Object} spec.control - Control config
-   * @param {Object[]} spec.variants - Array of { name: string, config: Object }
-   * @param {number} [spec.targetSampleSize=100] - Target observations per variant
-   * @param {string} [spec.description] - Human-readable description
-   * @returns {Object} Created experiment summary
-   */
-  function createExperiment(experimentId, spec) {
-    if (!experimentId || typeof experimentId !== 'string') {
-      throw new Error('experimentId must be a non-empty string');
-    }
-    if (experiments[experimentId]) {
-      throw new Error('Experiment already exists: ' + experimentId);
-    }
-    if (experimentOrder.length >= maxExperiments) {
-      throw new Error('Maximum experiments reached (' + maxExperiments + ')');
-    }
-    if (!spec || !spec.control) {
-      throw new Error('spec.control is required');
-    }
-    if (!spec.variants || !Array.isArray(spec.variants) || spec.variants.length === 0) {
-      throw new Error('spec.variants must be a non-empty array');
-    }
-    // Validate variant names are unique
-    var names = { control: true };
-    for (var i = 0; i < spec.variants.length; i++) {
-      var v = spec.variants[i];
-      if (!v.name || typeof v.name !== 'string') {
-        throw new Error('Each variant must have a name string');
-      }
-      if (names[v.name]) {
-        throw new Error('Duplicate variant name: ' + v.name);
-      }
-      names[v.name] = true;
-    }
-
-    var allVariants = [{ name: 'control', config: spec.control }];
-    for (var j = 0; j < spec.variants.length; j++) {
-      allVariants.push({ name: spec.variants[j].name, config: spec.variants[j].config || {} });
-    }
-
-    var variantData = {};
-    for (var k = 0; k < allVariants.length; k++) {
-      variantData[allVariants[k].name] = {
-        name: allVariants[k].name,
-        config: allVariants[k].config,
-        solves: 0,
-        fails: 0,
-        abandons: 0,
-        solveTimes: [],
-        users: {},
-      };
-    }
-
-    experiments[experimentId] = {
-      id: experimentId,
-      description: spec.description || '',
-      status: 'running',
-      createdAt: Date.now(),
-      endedAt: null,
-      targetSampleSize: spec.targetSampleSize > 0 ? spec.targetSampleSize : 100,
-      variantNames: allVariants.map(function (v) { return v.name; }),
-      variants: variantData,
-      winner: null,
-      userAssignments: {},
-    };
-    experimentOrder.push(experimentId);
-
-    return { id: experimentId, variants: allVariants.length, status: 'running' };
-  }
-
-  /**
-   * Deterministically assign a user to a variant.
-   * Subsequent calls with the same experimentId + userId return the same variant.
-   * @param {string} experimentId
-   * @param {string} userId
-   * @returns {{ variant: string, config: Object }}
-   */
-  function assignUser(experimentId, userId) {
-    var exp = experiments[experimentId];
-    if (!exp) throw new Error('Unknown experiment: ' + experimentId);
-    if (!userId || typeof userId !== 'string') throw new Error('userId must be a non-empty string');
-
-    // Check existing assignment
-    if (exp.userAssignments[userId]) {
-      var assignedName = exp.userAssignments[userId];
-      return { variant: assignedName, config: exp.variants[assignedName].config };
-    }
-
-    // Deterministic hash-based assignment
-    var hash = djb2(experimentId + ':' + userId);
-    var idx = hash % exp.variantNames.length;
-    var variantName = exp.variantNames[idx];
-    exp.userAssignments[userId] = variantName;
-
-    return { variant: variantName, config: exp.variants[variantName].config };
-  }
-
-  /**
-   * Record an event (solve, fail, abandon) for a user in an experiment.
-   * @param {string} experimentId
-   * @param {string} userId
-   * @param {string} eventType - 'solve' | 'fail' | 'abandon'
-   * @param {Object} [data]
-   * @param {number} [data.timeMs] - Solve time in milliseconds (for 'solve' events)
-   */
-  function recordEvent(experimentId, userId, eventType, data) {
-    var exp = experiments[experimentId];
-    if (!exp) throw new Error('Unknown experiment: ' + experimentId);
-    if (exp.status !== 'running') return; // Silently ignore if stopped
-
-    // Auto-assign if not already
-    var assignment = assignUser(experimentId, userId);
-    var variant = exp.variants[assignment.variant];
-    data = data || {};
-
-    switch (eventType) {
-      case 'solve':
-        variant.solves++;
-        if (typeof data.timeMs === 'number' && data.timeMs > 0) {
-          variant.solveTimes.push(data.timeMs);
-        }
-        break;
-      case 'fail':
-        variant.fails++;
-        break;
-      case 'abandon':
-        variant.abandons++;
-        break;
-      default:
-        throw new Error('Unknown event type: ' + eventType + '. Expected solve/fail/abandon');
-    }
-
-    // Track per-user events
-    if (!variant.users[userId]) variant.users[userId] = [];
-    variant.users[userId].push({ type: eventType, time: Date.now(), data: data });
-
-    // Check early stopping
-    if (earlyStoppingEnabled && exp.status === 'running') {
-      var ready = true;
-      for (var i = 0; i < exp.variantNames.length; i++) {
-        var vd = exp.variants[exp.variantNames[i]];
-        if (vd.solves + vd.fails + vd.abandons < defaultMinSample) {
-          ready = false;
-          break;
-        }
-      }
-      if (ready) {
-        var analysis = _analyzeInternal(exp);
-        if (analysis.significant && analysis.pValue <= earlyStoppingConfidence) {
-          exp.status = 'completed';
-          exp.endedAt = Date.now();
-          exp.winner = analysis.winner;
-          _fireCallbacks(experimentId, analysis);
-        }
-      }
-    }
-  }
-
-  /**
-   * Analyze an experiment's results.
-   * @param {string} experimentId
-   * @returns {Object} Analysis results with statistics
-   */
-  function analyzeExperiment(experimentId) {
-    var exp = experiments[experimentId];
-    if (!exp) throw new Error('Unknown experiment: ' + experimentId);
-    return _analyzeInternal(exp);
-  }
-
-  function _analyzeInternal(exp) {
-    var variantStats = [];
-    var chiVariants = [];
-    for (var i = 0; i < exp.variantNames.length; i++) {
-      var name = exp.variantNames[i];
-      var vd = exp.variants[name];
-      var stats = computeVariantStats(vd);
-      variantStats.push(stats);
-      chiVariants.push({ solves: vd.solves, attempts: stats.attempts });
-    }
-
-    // Chi-squared across all variants
-    var chiResult = chiSquaredTest(chiVariants);
-
-    // Pairwise z-tests: each variant vs control
-    var controlData = exp.variants['control'];
-    var controlAttempts = controlData.solves + controlData.fails;
-    var pairwise = [];
-    for (var j = 1; j < exp.variantNames.length; j++) {
-      var vName = exp.variantNames[j];
-      var vData = exp.variants[vName];
-      var vAttempts = vData.solves + vData.fails;
-      var zResult = proportionZTest(controlAttempts, controlData.solves, vAttempts, vData.solves);
-      var controlRate = controlAttempts > 0 ? controlData.solves / controlAttempts : 0;
-      var variantRate = vAttempts > 0 ? vData.solves / vAttempts : 0;
-      var lift = controlRate > 0 ? ((variantRate - controlRate) / controlRate) * 100 : 0;
-      pairwise.push({
-        variant: vName,
-        vsControl: {
-          z: Math.round(zResult.z * 1000) / 1000,
-          pValue: zResult.pValue,
-          significant: zResult.pValue < defaultSignificance,
-          lift: Math.round(lift * 100) / 100,
-        },
-      });
-    }
-
-    // Determine winner
-    var winner = null;
-    var bestRate = -1;
-    var significant = chiResult.pValue < defaultSignificance;
-
-    var sufficientData = true;
-    for (var m = 0; m < variantStats.length; m++) {
-      if (variantStats[m].attempts < defaultMinSample) {
-        sufficientData = false;
-        break;
-      }
-    }
-
-    if (sufficientData && significant) {
-      for (var k = 0; k < variantStats.length; k++) {
-        if (variantStats[k].solveRate > bestRate) {
-          bestRate = variantStats[k].solveRate;
-          winner = variantStats[k].name;
-        }
-      }
-    }
-
-    return {
-      experimentId: exp.id,
-      status: exp.status,
-      variants: variantStats,
-      chiSquared: {
-        value: Math.round(chiResult.chiSq * 1000) / 1000,
-        df: chiResult.df,
-        pValue: chiResult.pValue,
-      },
-      pairwiseTests: pairwise,
-      significant: significant,
-      pValue: chiResult.pValue,
-      sufficientData: sufficientData,
-      winner: winner,
-      recommendation: _getRecommendation(winner, significant, sufficientData, variantStats),
-    };
-  }
-
-  function _getRecommendation(winner, significant, sufficientData, variantStats) {
-    if (!sufficientData) {
-      var current = Infinity;
-      for (var i = 0; i < variantStats.length; i++) {
-        if (variantStats[i].attempts < current) current = variantStats[i].attempts;
-      }
-      return 'Insufficient data. Minimum ' + defaultMinSample + ' attempts per variant needed (lowest: ' + current + ').';
-    }
-    if (!significant) {
-      return 'No statistically significant difference detected. Continue collecting data or accept null hypothesis.';
-    }
-    if (winner === 'control') {
-      return 'Control performs best. No change recommended.';
-    }
-    return 'Variant "' + winner + '" is the winner. Consider adopting its configuration.';
-  }
-
-  /**
-   * Stop an experiment manually.
-   * @param {string} experimentId
-   * @returns {Object} Final analysis
-   */
-  function stopExperiment(experimentId) {
-    var exp = experiments[experimentId];
-    if (!exp) throw new Error('Unknown experiment: ' + experimentId);
-    exp.status = 'stopped';
-    exp.endedAt = Date.now();
-    var analysis = _analyzeInternal(exp);
-    exp.winner = analysis.winner;
-    return analysis;
-  }
-
-  /**
-   * Get experiment info.
-   * @param {string} experimentId
-   * @returns {Object|null}
-   */
-  function getExperiment(experimentId) {
-    var exp = experiments[experimentId];
-    if (!exp) return null;
-    return {
-      id: exp.id,
-      description: exp.description,
-      status: exp.status,
-      createdAt: exp.createdAt,
-      endedAt: exp.endedAt,
-      variantNames: exp.variantNames.slice(),
-      targetSampleSize: exp.targetSampleSize,
-      winner: exp.winner,
-      totalUsers: Object.keys(exp.userAssignments).length,
-    };
-  }
-
-  /**
-   * List all experiments.
-   * @param {Object} [filters]
-   * @param {string} [filters.status] - Filter by status (running/stopped/completed)
-   * @returns {Object[]}
-   */
-  function listExperiments(filters) {
-    filters = filters || {};
-    var results = [];
-    var order = experimentOrder.toArray();
-    for (var i = 0; i < order.length; i++) {
-      var exp = experiments[order[i]];
-      if (!exp) continue;
-      if (filters.status && exp.status !== filters.status) continue;
-      results.push(getExperiment(exp.id));
-    }
-    return results;
-  }
-
-  /**
-   * Delete an experiment.
-   * @param {string} experimentId
-   * @returns {boolean} True if deleted
-   */
-  function deleteExperiment(experimentId) {
-    if (!experiments[experimentId]) return false;
-    delete experiments[experimentId];
-    experimentOrder.remove(experimentId);
-    return true;
-  }
-
-  /**
-   * Get variant assignment counts for an experiment.
-   * @param {string} experimentId
-   * @returns {Object} Map of variant name to user count
-   */
-  function getAssignmentCounts(experimentId) {
-    var exp = experiments[experimentId];
-    if (!exp) throw new Error('Unknown experiment: ' + experimentId);
-    var counts = {};
-    for (var i = 0; i < exp.variantNames.length; i++) {
-      counts[exp.variantNames[i]] = 0;
-    }
-    var keys = Object.keys(exp.userAssignments);
-    for (var j = 0; j < keys.length; j++) {
-      var v2 = exp.userAssignments[keys[j]];
-      counts[v2] = (counts[v2] || 0) + 1;
-    }
-    return counts;
-  }
-
-  /**
-   * Register a callback for experiment completion (via early stopping).
-   * @param {Function} fn - Called with (experimentId, analysis)
-   */
-  function onResult(fn) {
-    if (typeof fn === 'function') onResultCallbacks.push(fn);
-  }
-
-  function _fireCallbacks(experimentId, analysis) {
-    for (var i = 0; i < onResultCallbacks.length; i++) {
-      try { onResultCallbacks[i](experimentId, analysis); } catch (e) { /* ignore */ }
-    }
-  }
-
-  /**
-   * Export all experiment data as JSON-serializable object.
-   * @returns {Object}
-   */
-  function exportState() {
-    var data = [];
-    var order = experimentOrder.toArray();
-    for (var i = 0; i < order.length; i++) {
-      var exp = experiments[order[i]];
-      if (!exp) continue;
-      var variantExport = {};
-      for (var j = 0; j < exp.variantNames.length; j++) {
-        var name = exp.variantNames[j];
-        var vd = exp.variants[name];
-        variantExport[name] = {
-          name: vd.name,
-          config: vd.config,
-          solves: vd.solves,
-          fails: vd.fails,
-          abandons: vd.abandons,
-          solveTimes: vd.solveTimes.slice(),
-        };
-      }
-      data.push({
-        id: exp.id,
-        description: exp.description,
-        status: exp.status,
-        createdAt: exp.createdAt,
-        endedAt: exp.endedAt,
-        targetSampleSize: exp.targetSampleSize,
-        variantNames: exp.variantNames.slice(),
-        variants: variantExport,
-        winner: exp.winner,
-      });
-    }
-    return { experiments: data, exportedAt: Date.now() };
-  }
-
-  /**
-   * Import experiment data from a previous export.
-   * @param {Object} state - Data from exportState()
-   * @returns {number} Number of experiments imported
-   */
-  function importState(state) {
-    if (!state || !Array.isArray(state.experiments)) return 0;
-    var imported = 0;
-    for (var i = 0; i < state.experiments.length; i++) {
-      var e = state.experiments[i];
-      if (!e.id || experiments[e.id]) continue;
-      if (experimentOrder.length >= maxExperiments) break;
-
-      var variantData = {};
-      for (var j = 0; j < e.variantNames.length; j++) {
-        var name = e.variantNames[j];
-        var src = (e.variants && e.variants[name]) || {};
-        variantData[name] = {
-          name: name,
-          config: src.config || {},
-          solves: src.solves || 0,
-          fails: src.fails || 0,
-          abandons: src.abandons || 0,
-          solveTimes: Array.isArray(src.solveTimes) ? src.solveTimes.slice() : [],
-          users: {},
-        };
-      }
-
-      experiments[e.id] = {
-        id: e.id,
-        description: e.description || '',
-        status: e.status || 'stopped',
-        createdAt: e.createdAt || Date.now(),
-        endedAt: e.endedAt || null,
-        targetSampleSize: e.targetSampleSize || 100,
-        variantNames: e.variantNames.slice(),
-        variants: variantData,
-        winner: e.winner || null,
-        userAssignments: {},
-      };
-      experimentOrder.push(e.id);
-      imported++;
-    }
-    return imported;
-  }
-
-  /**
-   * Generate a text summary of an experiment's results.
-   * @param {string} experimentId
-   * @returns {string}
-   */
-  function textReport(experimentId) {
-    var analysis = analyzeExperiment(experimentId);
-    var exp = experiments[experimentId];
-    var lines = [];
-    lines.push('A/B Experiment: ' + analysis.experimentId);
-    if (exp.description) lines.push('Description: ' + exp.description);
-    lines.push('Status: ' + analysis.status);
-    lines.push('');
-    lines.push('Variants:');
-    for (var i = 0; i < analysis.variants.length; i++) {
-      var v = analysis.variants[i];
-      lines.push('  ' + v.name + ':');
-      lines.push('    Attempts: ' + v.attempts);
-      lines.push('    Solve rate: ' + (v.solveRate * 100).toFixed(1) + '%');
-      lines.push('    Abandon rate: ' + (v.abandonRate * 100).toFixed(1) + '%');
-      if (v.avgSolveTimeMs > 0) {
-        lines.push('    Avg solve time: ' + v.avgSolveTimeMs + 'ms');
-        lines.push('    Median solve time: ' + v.medianSolveTimeMs + 'ms');
-        lines.push('    P95 solve time: ' + v.p95SolveTimeMs + 'ms');
-      }
-    }
-    lines.push('');
-    lines.push('Chi-squared: ' + analysis.chiSquared.value + ' (df=' + analysis.chiSquared.df + ', p=' + analysis.pValue.toFixed(4) + ')');
-    lines.push('Significant: ' + (analysis.significant ? 'Yes' : 'No'));
-    if (analysis.pairwiseTests.length > 0) {
-      lines.push('');
-      lines.push('Pairwise vs Control:');
-      for (var j = 0; j < analysis.pairwiseTests.length; j++) {
-        var pw = analysis.pairwiseTests[j];
-        lines.push('  ' + pw.variant + ': z=' + pw.vsControl.z + ', p=' + pw.vsControl.pValue.toFixed(4) +
-          ', lift=' + pw.vsControl.lift + '%' +
-          (pw.vsControl.significant ? ' *' : ''));
-      }
-    }
-    lines.push('');
-    lines.push('Winner: ' + (analysis.winner || 'None'));
-    lines.push('Recommendation: ' + analysis.recommendation);
-    return lines.join('\n');
-  }
-
-  return {
-    createExperiment: createExperiment,
-    assignUser: assignUser,
-    recordEvent: recordEvent,
-    analyzeExperiment: analyzeExperiment,
-    stopExperiment: stopExperiment,
-    getExperiment: getExperiment,
-    listExperiments: listExperiments,
-    deleteExperiment: deleteExperiment,
-    getAssignmentCounts: getAssignmentCounts,
-    onResult: onResult,
-    exportState: exportState,
-    importState: importState,
-    textReport: textReport,
-  };
-}
-
 // ─── Fraud Ring Detector ──────────────────────────────────────────────────────
 
 /**
@@ -9251,757 +8589,6 @@ function createFraudRingDetector(options) {
     getStats: getStats, findTimingClusters: findTimingClusters, findSharedFingerprints: findSharedFingerprints,
     findIPClusters: findIPClusters, findSequentialPatterns: findSequentialPatterns, computeRingScore: computeRingScore,
     exportState: exportState, importState: importState, generateReport: generateReport, reset: reset
-  };
-}
-
-// ── Compliance Report Generator ───────────────────────────────────────
-
-/**
- * Creates a compliance report generator that evaluates CAPTCHA system
- * configuration and runtime metrics against accessibility, security,
- * privacy, and operational standards.
- *
- * Checks against:
- * - WCAG 2.1 AA accessibility guidelines
- * - GDPR / privacy data-retention requirements
- * - OWASP bot-mitigation best practices
- * - Operational health thresholds
- *
- * @param {Object} [options] - Generator configuration
- * @param {string} [options.systemName="gif-captcha"] - System identifier for reports
- * @param {number} [options.maxDataRetentionDays=30] - GDPR data retention limit in days
- * @param {number} [options.minSolveRatePercent=70] - Minimum acceptable human solve rate
- * @param {number} [options.maxSolveTimeMs=60000] - Maximum acceptable solve time
- * @param {number} [options.maxFailRatePercent=50] - Maximum acceptable failure rate
- * @param {number} [options.minBotBlockRatePercent=90] - Minimum bot detection rate
- * @returns {Object} Compliance report generator instance
- */
-/**
- * Count severity occurrences across an array of findings.
- * @param {Object[]} findings - Array of finding objects with .severity
- * @param {Object} SEV - Severity constants (PASS, CRITICAL, WARNING, INFO)
- * @returns {{ passed: number, criticals: number, warnings: number, infos: number }}
- */
-function _countSeverities(findings, SEV) {
-  var passed = 0, criticals = 0, warnings = 0, infos = 0;
-  for (var i = 0; i < findings.length; i++) {
-    var s = findings[i].severity;
-    if (s === SEV.PASS) passed++;
-    else if (s === SEV.CRITICAL) criticals++;
-    else if (s === SEV.WARNING) warnings++;
-    else if (s === SEV.INFO) infos++;
-  }
-  return { passed: passed, criticals: criticals, warnings: warnings, infos: infos };
-}
-
-/**
- * Compute per-category compliance scores from findings.
- * @param {Object[]} findings - Array of finding objects with .category and .severity
- * @param {Object} SEV - Severity constants
- * @returns {Object} Map of category -> { score, total, passed, criticals, warnings }
- */
-function _computeAllCategoryScores(findings, SEV) {
-  var buckets = {};
-  for (var i = 0; i < findings.length; i++) {
-    var cat = findings[i].category;
-    if (!buckets[cat]) buckets[cat] = [];
-    buckets[cat].push(findings[i]);
-  }
-  var scores = {};
-  var cats = Object.keys(buckets);
-  for (var c = 0; c < cats.length; c++) {
-    var catName = cats[c];
-    var items = buckets[catName];
-    var counts = _countSeverities(items, SEV);
-    var scorable = items.length - counts.infos;
-    var score = scorable > 0 ? Math.round((counts.passed / scorable) * 100) : 100;
-    scores[catName] = {
-      score: score,
-      total: items.length,
-      passed: counts.passed,
-      criticals: counts.criticals,
-      warnings: counts.warnings
-    };
-  }
-  return scores;
-}
-
-/**
- * Compute a weighted average score from category scores and weight map.
- * @param {Object} categoryScores - Map of category -> { score }
- * @param {Object} weights - Map of category -> weight (number)
- * @returns {number} Weighted average (0-100), rounded
- */
-function _weightedAverage(categoryScores, weights) {
-  var total = 0, weightSum = 0;
-  var cats = Object.keys(weights);
-  for (var i = 0; i < cats.length; i++) {
-    var cat = cats[i];
-    if (categoryScores[cat]) {
-      total += categoryScores[cat].score * weights[cat];
-      weightSum += weights[cat];
-    }
-  }
-  return weightSum > 0 ? Math.round(total / weightSum) : 0;
-}
-
-function createComplianceReporter(options) {
-  options = options || {};
-  var systemName = options.systemName || "gif-captcha";
-  var maxRetentionDays = options.maxDataRetentionDays > 0 ? options.maxDataRetentionDays : 30;
-  var minSolveRate = typeof options.minSolveRatePercent === "number" ? options.minSolveRatePercent : 70;
-  var maxSolveTimeMs = options.maxSolveTimeMs > 0 ? options.maxSolveTimeMs : 60000;
-  var maxFailRate = typeof options.maxFailRatePercent === "number" ? options.maxFailRatePercent : 50;
-  var minBotBlockRate = typeof options.minBotBlockRatePercent === "number" ? options.minBotBlockRatePercent : 90;
-
-  var SEVERITY = { CRITICAL: "critical", WARNING: "warning", INFO: "info", PASS: "pass" };
-  var CATEGORY = {
-    ACCESSIBILITY: "accessibility",
-    PRIVACY: "privacy",
-    SECURITY: "security",
-    OPERATIONAL: "operational"
-  };
-
-  /**
-   * Run all compliance checks against provided configuration and metrics.
-   *
-   * @param {Object} config - CAPTCHA system configuration
-   * @param {boolean} [config.audioAlternative] - Whether audio CAPTCHA is available
-   * @param {boolean} [config.keyboardNavigable] - Whether CAPTCHA is keyboard-navigable
-   * @param {string} [config.ariaLabel] - ARIA label for the CAPTCHA element
-   * @param {number} [config.colorContrast] - Color contrast ratio (e.g. 4.5)
-   * @param {number} [config.timeLimitMs] - Time limit given to solve
-   * @param {boolean} [config.canExtendTime] - Whether user can extend time
-   * @param {string[]} [config.supportedLanguages] - List of supported locale codes
-   * @param {number} [config.dataRetentionDays] - How long data is retained
-   * @param {boolean} [config.consentRequired] - Whether consent is collected
-   * @param {boolean} [config.anonymization] - Whether data is anonymised
-   * @param {boolean} [config.deletionSupported] - Whether right-to-delete is supported
-   * @param {boolean} [config.rateLimitEnabled] - Whether rate limiting is active
-   * @param {boolean} [config.tokenSigned] - Whether tokens use HMAC signing
-   * @param {boolean} [config.httpsOnly] - Whether HTTPS is enforced
-   * @param {boolean} [config.inputSanitized] - Whether input is sanitised
-   * @param {number} [config.maxAttempts] - Max attempts before lockout
-   * @param {boolean} [config.replayProtection] - Whether replay attacks are prevented
-   * @param {Object} [metrics] - Runtime metrics snapshot
-   * @param {number} [metrics.totalChallenges] - Total challenges served
-   * @param {number} [metrics.totalSolves] - Successful solves
-   * @param {number} [metrics.totalFailures] - Failed attempts
-   * @param {number} [metrics.avgSolveTimeMs] - Average solve time in ms
-   * @param {number} [metrics.p95SolveTimeMs] - P95 solve time in ms
-   * @param {number} [metrics.botAttempts] - Detected bot attempts
-   * @param {number} [metrics.botBlocked] - Blocked bot attempts
-   * @param {number} [metrics.uptimePercent] - System uptime percentage
-   * @param {number} [metrics.avgResponseTimeMs] - Average server response time
-   * @param {number} [metrics.errorCount] - Server error count
-   * @returns {Object} Compliance report
-   */
-  function generateReport(config, metrics) {
-    config = config || {};
-    metrics = metrics || {};
-    var findings = [];
-    var now = new Date();
-
-    // ── Accessibility Checks (WCAG 2.1 AA) ─────────────────────────
-
-    findings.push({
-      id: "ACC-001",
-      category: CATEGORY.ACCESSIBILITY,
-      title: "Audio alternative available",
-      description: "WCAG 1.1.1: Non-text content must have a text or audio alternative",
-      severity: config.audioAlternative ? SEVERITY.PASS : SEVERITY.CRITICAL,
-      recommendation: config.audioAlternative ? null : "Enable audio CAPTCHA alternative for visually impaired users"
-    });
-
-    findings.push({
-      id: "ACC-002",
-      category: CATEGORY.ACCESSIBILITY,
-      title: "Keyboard navigation support",
-      description: "WCAG 2.1.1: All functionality must be operable through a keyboard",
-      severity: config.keyboardNavigable ? SEVERITY.PASS : SEVERITY.CRITICAL,
-      recommendation: config.keyboardNavigable ? null : "Ensure CAPTCHA can be completed using keyboard only"
-    });
-
-    findings.push({
-      id: "ACC-003",
-      category: CATEGORY.ACCESSIBILITY,
-      title: "ARIA labelling",
-      description: "WCAG 4.1.2: UI components must have accessible names and roles",
-      severity: config.ariaLabel ? SEVERITY.PASS : SEVERITY.WARNING,
-      recommendation: config.ariaLabel ? null : "Add aria-label or aria-labelledby to the CAPTCHA container"
-    });
-
-    var contrast = typeof config.colorContrast === "number" ? config.colorContrast : 0;
-    findings.push({
-      id: "ACC-004",
-      category: CATEGORY.ACCESSIBILITY,
-      title: "Color contrast ratio",
-      description: "WCAG 1.4.3: Text must have a contrast ratio of at least 4.5:1",
-      severity: contrast >= 4.5 ? SEVERITY.PASS : contrast >= 3 ? SEVERITY.WARNING : SEVERITY.CRITICAL,
-      recommendation: contrast >= 4.5 ? null : "Increase color contrast to at least 4.5:1 (current: " + contrast.toFixed(1) + ":1)"
-    });
-
-    var hasTimeLimit = typeof config.timeLimitMs === "number" && config.timeLimitMs > 0;
-    findings.push({
-      id: "ACC-005",
-      category: CATEGORY.ACCESSIBILITY,
-      title: "Time limit accommodation",
-      description: "WCAG 2.2.1: Users must be able to turn off, adjust, or extend time limits",
-      severity: !hasTimeLimit || config.canExtendTime ? SEVERITY.PASS : SEVERITY.WARNING,
-      recommendation: hasTimeLimit && !config.canExtendTime ? "Allow users to request additional time to complete the CAPTCHA" : null
-    });
-
-    var langCount = Array.isArray(config.supportedLanguages) ? config.supportedLanguages.length : 0;
-    findings.push({
-      id: "ACC-006",
-      category: CATEGORY.ACCESSIBILITY,
-      title: "Multilingual support",
-      description: "WCAG 3.1.1: Default language must be programmatically determinable",
-      severity: langCount >= 3 ? SEVERITY.PASS : langCount >= 1 ? SEVERITY.INFO : SEVERITY.WARNING,
-      recommendation: langCount < 3 ? "Support at least 3 languages for broader accessibility (" + langCount + " currently configured)" : null
-    });
-
-    // ── Privacy Checks (GDPR) ───────────────────────────────────────
-
-    var retDays = typeof config.dataRetentionDays === "number" ? config.dataRetentionDays : -1;
-    findings.push({
-      id: "PRV-001",
-      category: CATEGORY.PRIVACY,
-      title: "Data retention policy",
-      description: "GDPR Art. 5(1)(e): Data must not be kept longer than necessary",
-      severity: retDays >= 0 && retDays <= maxRetentionDays ? SEVERITY.PASS : retDays < 0 ? SEVERITY.CRITICAL : SEVERITY.WARNING,
-      recommendation: retDays < 0 ? "Define a data retention period (max " + maxRetentionDays + " days recommended)"
-        : retDays > maxRetentionDays ? "Reduce retention from " + retDays + " to " + maxRetentionDays + " days or less" : null
-    });
-
-    findings.push({
-      id: "PRV-002",
-      category: CATEGORY.PRIVACY,
-      title: "User consent collection",
-      description: "GDPR Art. 6: Processing requires a lawful basis (consent for CAPTCHAs)",
-      severity: config.consentRequired ? SEVERITY.PASS : SEVERITY.WARNING,
-      recommendation: config.consentRequired ? null : "Collect explicit consent before processing CAPTCHA interaction data"
-    });
-
-    findings.push({
-      id: "PRV-003",
-      category: CATEGORY.PRIVACY,
-      title: "Data anonymisation",
-      description: "GDPR Art. 25: Data protection by design and default",
-      severity: config.anonymization ? SEVERITY.PASS : SEVERITY.WARNING,
-      recommendation: config.anonymization ? null : "Anonymise or pseudonymise stored CAPTCHA interaction data"
-    });
-
-    findings.push({
-      id: "PRV-004",
-      category: CATEGORY.PRIVACY,
-      title: "Right to deletion",
-      description: "GDPR Art. 17: Users have the right to erasure of personal data",
-      severity: config.deletionSupported ? SEVERITY.PASS : SEVERITY.CRITICAL,
-      recommendation: config.deletionSupported ? null : "Implement data deletion endpoint for GDPR Art. 17 compliance"
-    });
-
-    // ── Security Checks (OWASP) ─────────────────────────────────────
-
-    findings.push({
-      id: "SEC-001",
-      category: CATEGORY.SECURITY,
-      title: "Rate limiting enabled",
-      description: "OWASP: Implement rate limiting to prevent automated attacks",
-      severity: config.rateLimitEnabled ? SEVERITY.PASS : SEVERITY.CRITICAL,
-      recommendation: config.rateLimitEnabled ? null : "Enable request rate limiting to prevent brute-force attacks"
-    });
-
-    findings.push({
-      id: "SEC-002",
-      category: CATEGORY.SECURITY,
-      title: "Token signing (HMAC)",
-      description: "OWASP: Validate CAPTCHA tokens server-side with cryptographic signatures",
-      severity: config.tokenSigned ? SEVERITY.PASS : SEVERITY.CRITICAL,
-      recommendation: config.tokenSigned ? null : "Use HMAC-signed tokens for stateless CAPTCHA validation"
-    });
-
-    findings.push({
-      id: "SEC-003",
-      category: CATEGORY.SECURITY,
-      title: "HTTPS enforcement",
-      description: "OWASP: Encrypt all CAPTCHA traffic to prevent interception",
-      severity: config.httpsOnly ? SEVERITY.PASS : SEVERITY.CRITICAL,
-      recommendation: config.httpsOnly ? null : "Enforce HTTPS for all CAPTCHA endpoints"
-    });
-
-    findings.push({
-      id: "SEC-004",
-      category: CATEGORY.SECURITY,
-      title: "Input sanitisation",
-      description: "OWASP: Sanitise all user input to prevent injection attacks",
-      severity: config.inputSanitized ? SEVERITY.PASS : SEVERITY.WARNING,
-      recommendation: config.inputSanitized ? null : "Sanitise CAPTCHA answer input before processing"
-    });
-
-    var maxAttempts = typeof config.maxAttempts === "number" ? config.maxAttempts : 0;
-    findings.push({
-      id: "SEC-005",
-      category: CATEGORY.SECURITY,
-      title: "Attempt limiting / lockout",
-      description: "OWASP: Lock out after repeated failures to prevent brute force",
-      severity: maxAttempts > 0 && maxAttempts <= 10 ? SEVERITY.PASS : maxAttempts > 10 ? SEVERITY.WARNING : SEVERITY.CRITICAL,
-      recommendation: maxAttempts <= 0 ? "Configure a maximum attempt limit (recommended: 3-5)"
-        : maxAttempts > 10 ? "Reduce max attempts from " + maxAttempts + " to 5 or fewer" : null
-    });
-
-    findings.push({
-      id: "SEC-006",
-      category: CATEGORY.SECURITY,
-      title: "Replay protection",
-      description: "Prevent reuse of solved CAPTCHA tokens",
-      severity: config.replayProtection ? SEVERITY.PASS : SEVERITY.WARNING,
-      recommendation: config.replayProtection ? null : "Implement one-time-use tokens to prevent replay attacks"
-    });
-
-    var botBlockRate = 0;
-    if (metrics.botAttempts > 0) {
-      botBlockRate = (metrics.botBlocked / metrics.botAttempts) * 100;
-    }
-    findings.push({
-      id: "SEC-007",
-      category: CATEGORY.SECURITY,
-      title: "Bot detection effectiveness",
-      description: "Bot block rate should be at least " + minBotBlockRate + "%",
-      severity: metrics.botAttempts === 0 ? SEVERITY.INFO
-        : botBlockRate >= minBotBlockRate ? SEVERITY.PASS
-        : botBlockRate >= minBotBlockRate * 0.8 ? SEVERITY.WARNING : SEVERITY.CRITICAL,
-      recommendation: metrics.botAttempts > 0 && botBlockRate < minBotBlockRate
-        ? "Bot block rate is " + botBlockRate.toFixed(1) + "% (target: " + minBotBlockRate + "%)" : null
-    });
-
-    // ── Operational Checks ──────────────────────────────────────────
-
-    var solveRate = 0;
-    if (metrics.totalChallenges > 0) {
-      solveRate = (metrics.totalSolves / metrics.totalChallenges) * 100;
-    }
-    findings.push({
-      id: "OPS-001",
-      category: CATEGORY.OPERATIONAL,
-      title: "Human solve rate",
-      description: "Solve rate should be at least " + minSolveRate + "% to avoid user frustration",
-      severity: metrics.totalChallenges === 0 ? SEVERITY.INFO
-        : solveRate >= minSolveRate ? SEVERITY.PASS
-        : solveRate >= minSolveRate * 0.8 ? SEVERITY.WARNING : SEVERITY.CRITICAL,
-      recommendation: metrics.totalChallenges > 0 && solveRate < minSolveRate
-        ? "Solve rate is " + solveRate.toFixed(1) + "% (target: " + minSolveRate + "%). Consider reducing difficulty." : null
-    });
-
-    var failRate = 0;
-    if (metrics.totalChallenges > 0) {
-      failRate = (metrics.totalFailures / metrics.totalChallenges) * 100;
-    }
-    findings.push({
-      id: "OPS-002",
-      category: CATEGORY.OPERATIONAL,
-      title: "Failure rate within threshold",
-      description: "Failure rate should be below " + maxFailRate + "%",
-      severity: metrics.totalChallenges === 0 ? SEVERITY.INFO
-        : failRate <= maxFailRate ? SEVERITY.PASS
-        : failRate <= maxFailRate * 1.2 ? SEVERITY.WARNING : SEVERITY.CRITICAL,
-      recommendation: metrics.totalChallenges > 0 && failRate > maxFailRate
-        ? "Failure rate is " + failRate.toFixed(1) + "% (threshold: " + maxFailRate + "%)" : null
-    });
-
-    var avgTime = typeof metrics.avgSolveTimeMs === "number" ? metrics.avgSolveTimeMs : 0;
-    findings.push({
-      id: "OPS-003",
-      category: CATEGORY.OPERATIONAL,
-      title: "Average solve time acceptable",
-      description: "Solve time should be under " + (maxSolveTimeMs / 1000) + "s",
-      severity: avgTime <= 0 ? SEVERITY.INFO
-        : avgTime <= maxSolveTimeMs ? SEVERITY.PASS
-        : avgTime <= maxSolveTimeMs * 1.5 ? SEVERITY.WARNING : SEVERITY.CRITICAL,
-      recommendation: avgTime > maxSolveTimeMs
-        ? "Average solve time is " + (avgTime / 1000).toFixed(1) + "s (target: " + (maxSolveTimeMs / 1000) + "s)" : null
-    });
-
-    var p95Time = typeof metrics.p95SolveTimeMs === "number" ? metrics.p95SolveTimeMs : 0;
-    findings.push({
-      id: "OPS-004",
-      category: CATEGORY.OPERATIONAL,
-      title: "P95 solve time acceptable",
-      description: "95th percentile solve time should be under " + (maxSolveTimeMs * 2 / 1000) + "s",
-      severity: p95Time <= 0 ? SEVERITY.INFO
-        : p95Time <= maxSolveTimeMs * 2 ? SEVERITY.PASS
-        : p95Time <= maxSolveTimeMs * 3 ? SEVERITY.WARNING : SEVERITY.CRITICAL,
-      recommendation: p95Time > maxSolveTimeMs * 2
-        ? "P95 solve time is " + (p95Time / 1000).toFixed(1) + "s — consider simplifying hard challenges" : null
-    });
-
-    var uptime = typeof metrics.uptimePercent === "number" ? metrics.uptimePercent : 100;
-    findings.push({
-      id: "OPS-005",
-      category: CATEGORY.OPERATIONAL,
-      title: "System availability",
-      description: "Uptime should be at least 99.5%",
-      severity: uptime >= 99.5 ? SEVERITY.PASS : uptime >= 99 ? SEVERITY.WARNING : SEVERITY.CRITICAL,
-      recommendation: uptime < 99.5 ? "Uptime is " + uptime.toFixed(2) + "% (target: 99.5%)" : null
-    });
-
-    var responseTime = typeof metrics.avgResponseTimeMs === "number" ? metrics.avgResponseTimeMs : 0;
-    findings.push({
-      id: "OPS-006",
-      category: CATEGORY.OPERATIONAL,
-      title: "Server response time",
-      description: "Average response time should be under 500ms",
-      severity: responseTime <= 0 ? SEVERITY.INFO
-        : responseTime <= 500 ? SEVERITY.PASS
-        : responseTime <= 1000 ? SEVERITY.WARNING : SEVERITY.CRITICAL,
-      recommendation: responseTime > 500 ? "Average response time is " + responseTime + "ms (target: <500ms)" : null
-    });
-
-    var errCount = typeof metrics.errorCount === "number" ? metrics.errorCount : 0;
-    var errRate = metrics.totalChallenges > 0 ? (errCount / metrics.totalChallenges) * 100 : 0;
-    findings.push({
-      id: "OPS-007",
-      category: CATEGORY.OPERATIONAL,
-      title: "Error rate",
-      description: "Server error rate should be below 1%",
-      severity: metrics.totalChallenges === 0 ? SEVERITY.INFO
-        : errRate <= 1 ? SEVERITY.PASS
-        : errRate <= 5 ? SEVERITY.WARNING : SEVERITY.CRITICAL,
-      recommendation: errRate > 1 ? "Error rate is " + errRate.toFixed(2) + "% (" + errCount + " errors)" : null
-    });
-
-    // ── Compute Scores ──────────────────────────────────────────────
-
-    var categoryScores = _computeAllCategoryScores(findings, SEVERITY);
-
-    // Overall score: weighted average (security 35%, privacy 25%, accessibility 25%, operational 15%)
-    var weights = { security: 35, privacy: 25, accessibility: 25, operational: 15 };
-    var overallScore = _weightedAverage(categoryScores, weights);
-    var grade = overallScore >= 90 ? "A" : overallScore >= 80 ? "B" : overallScore >= 70 ? "C" : overallScore >= 60 ? "D" : "F";
-
-    var counts = _countSeverities(findings, SEVERITY);
-    var totalCriticals = counts.criticals;
-    var totalWarnings = counts.warnings;
-    var totalPassed = counts.passed;
-
-    return {
-      system: systemName,
-      generatedAt: now.toISOString(),
-      overallScore: overallScore,
-      grade: grade,
-      totalFindings: findings.length,
-      passed: totalPassed,
-      criticals: totalCriticals,
-      warnings: totalWarnings,
-      categoryScores: categoryScores,
-      findings: findings
-    };
-  }
-
-  /**
-   * Generate a minimal configuration template that would achieve a passing score.
-   * Useful for bootstrapping new CAPTCHA deployments.
-   *
-   * @returns {Object} Recommended configuration object
-   */
-  function getRecommendedConfig() {
-    return {
-      audioAlternative: true,
-      keyboardNavigable: true,
-      ariaLabel: "Security verification",
-      colorContrast: 4.5,
-      timeLimitMs: 120000,
-      canExtendTime: true,
-      supportedLanguages: ["en", "es", "fr"],
-      dataRetentionDays: maxRetentionDays,
-      consentRequired: true,
-      anonymization: true,
-      deletionSupported: true,
-      rateLimitEnabled: true,
-      tokenSigned: true,
-      httpsOnly: true,
-      inputSanitized: true,
-      maxAttempts: 5,
-      replayProtection: true
-    };
-  }
-
-  /**
-   * Compare two reports and return a diff summary showing improvements
-   * and regressions between audits.
-   *
-   * @param {Object} oldReport - Previous compliance report
-   * @param {Object} newReport - Current compliance report
-   * @returns {Object} Diff summary with improved, regressed, and unchanged counts
-   */
-  function compareReports(oldReport, newReport) {
-    if (!oldReport || !newReport || !oldReport.findings || !newReport.findings) {
-      return { error: "invalid_reports", improved: 0, regressed: 0, unchanged: 0, details: [] };
-    }
-    var oldMap = {};
-    for (var oi = 0; oi < oldReport.findings.length; oi++) {
-      oldMap[oldReport.findings[oi].id] = oldReport.findings[oi].severity;
-    }
-
-    var severityRank = {};
-    severityRank[SEVERITY.PASS] = 0;
-    severityRank[SEVERITY.INFO] = 1;
-    severityRank[SEVERITY.WARNING] = 2;
-    severityRank[SEVERITY.CRITICAL] = 3;
-
-    var improved = 0, regressed = 0, unchanged = 0;
-    var details = [];
-    for (var ni = 0; ni < newReport.findings.length; ni++) {
-      var finding = newReport.findings[ni];
-      var oldSev = oldMap[finding.id];
-      if (oldSev === undefined) {
-        details.push({ id: finding.id, change: "new", severity: finding.severity });
-        continue;
-      }
-      var oldRank = severityRank[oldSev] !== undefined ? severityRank[oldSev] : 1;
-      var newRank = severityRank[finding.severity] !== undefined ? severityRank[finding.severity] : 1;
-      if (newRank < oldRank) {
-        improved++;
-        details.push({ id: finding.id, change: "improved", from: oldSev, to: finding.severity });
-      } else if (newRank > oldRank) {
-        regressed++;
-        details.push({ id: finding.id, change: "regressed", from: oldSev, to: finding.severity });
-      } else {
-        unchanged++;
-      }
-    }
-
-    return {
-      oldScore: oldReport.overallScore,
-      newScore: newReport.overallScore,
-      scoreDelta: newReport.overallScore - oldReport.overallScore,
-      improved: improved,
-      regressed: regressed,
-      unchanged: unchanged,
-      details: details
-    };
-  }
-
-  /**
-   * Format a report as a plain-text summary suitable for terminal output or logging.
-   *
-   * @param {Object} report - Compliance report from generateReport()
-   * @returns {string} Formatted text report
-   */
-  function formatReportText(report) {
-    if (!report) return "";
-    var lines = [];
-    lines.push("=== CAPTCHA Compliance Report ===");
-    lines.push("System: " + report.system);
-    lines.push("Generated: " + report.generatedAt);
-    lines.push("Overall Score: " + report.overallScore + "/100 (Grade: " + report.grade + ")");
-    lines.push("");
-
-    var cats = ["accessibility", "privacy", "security", "operational"];
-    for (var ci = 0; ci < cats.length; ci++) {
-      var cat = cats[ci];
-      var cs = report.categoryScores[cat];
-      if (!cs) continue;
-      lines.push("  " + cat.charAt(0).toUpperCase() + cat.slice(1) + ": " + cs.score + "/100 (" + cs.passed + "/" + cs.total + " passed)");
-    }
-    lines.push("");
-
-    var actionItems = [];
-    for (var fi = 0; fi < report.findings.length; fi++) {
-      var f = report.findings[fi];
-      if (f.severity === SEVERITY.CRITICAL || f.severity === SEVERITY.WARNING) {
-        actionItems.push(f);
-      }
-    }
-
-    if (actionItems.length > 0) {
-      lines.push("Action Items:");
-      for (var ai = 0; ai < actionItems.length; ai++) {
-        var item = actionItems[ai];
-        var icon = item.severity === SEVERITY.CRITICAL ? "[CRITICAL]" : "[WARNING]";
-        lines.push("  " + icon + " " + item.id + " " + item.title);
-        if (item.recommendation) {
-          lines.push("    -> " + item.recommendation);
-        }
-      }
-    } else {
-      lines.push("No action items - all checks passed!");
-    }
-
-    return lines.join("\n");
-  }
-
-  /**
-   * Render a compliance report as a self-contained HTML page with inline
-   * CSS styling, color-coded severity badges, category score bars, and
-   * a summary dashboard. The output is a complete HTML document that can
-   * be saved to a file and opened in any browser.
-   *
-   * @param {Object} report - Compliance report from generateReport()
-   * @param {Object} [htmlOptions] - Rendering options
-   * @param {string} [htmlOptions.title] - Custom page title
-   * @param {boolean} [htmlOptions.darkMode=false] - Use dark colour scheme
-   * @param {boolean} [htmlOptions.includeTimestamp=true] - Show generation timestamp
-   * @returns {string} Complete HTML document string
-   */
-  function formatReportHtml(report, htmlOptions) {
-    if (!report) return "";
-    htmlOptions = htmlOptions || {};
-    var title = htmlOptions.title || "CAPTCHA Compliance Report — " + (report.system || "System");
-    var dark = !!htmlOptions.darkMode;
-    var showTime = htmlOptions.includeTimestamp !== false;
-
-    // Colour palette
-    var bg = dark ? "#1a1a2e" : "#f8f9fa";
-    var cardBg = dark ? "#16213e" : "#ffffff";
-    var textColor = dark ? "#e0e0e0" : "#333333";
-    var mutedText = dark ? "#a0a0a0" : "#6c757d";
-    var borderColor = dark ? "#2a2a4a" : "#dee2e6";
-
-    var severityColors = {
-      critical: { bg: "#dc3545", text: "#fff" },
-      warning:  { bg: "#ffc107", text: "#333" },
-      info:     { bg: "#17a2b8", text: "#fff" },
-      pass:     { bg: "#28a745", text: "#fff" }
-    };
-
-    var gradeColors = {
-      A: "#28a745", B: "#5cb85c", C: "#ffc107", D: "#fd7e14", F: "#dc3545"
-    };
-
-    var gradeColor = gradeColors[report.grade] || "#6c757d";
-
-    // Build category score bars
-    var cats = ["accessibility", "privacy", "security", "operational"];
-    var catLabels = { accessibility: "Accessibility", privacy: "Privacy", security: "Security", operational: "Operational" };
-    var catIcons = { accessibility: "♿", privacy: "🔒", security: "🛡️", operational: "⚙️" };
-
-    var scoreBarsHtml = "";
-    for (var ci = 0; ci < cats.length; ci++) {
-      var cat = cats[ci];
-      var cs = report.categoryScores[cat];
-      if (!cs) continue;
-      var barColor = cs.score >= 80 ? "#28a745" : cs.score >= 60 ? "#ffc107" : "#dc3545";
-      scoreBarsHtml += '<div style="margin-bottom:12px">'
-        + '<div style="display:flex;justify-content:space-between;margin-bottom:4px">'
-        + '<span>' + catIcons[cat] + ' ' + catLabels[cat] + '</span>'
-        + '<span style="font-weight:600">' + cs.score + '/100</span>'
-        + '</div>'
-        + '<div style="background:' + borderColor + ';border-radius:8px;height:10px;overflow:hidden">'
-        + '<div style="width:' + cs.score + '%;height:100%;background:' + barColor + ';border-radius:8px;transition:width 0.5s"></div>'
-        + '</div>'
-        + '<div style="font-size:12px;color:' + mutedText + ';margin-top:2px">'
-        + cs.passed + '/' + cs.total + ' passed'
-        + (cs.criticals > 0 ? ' · ' + cs.criticals + ' critical' : '')
-        + (cs.warnings > 0 ? ' · ' + cs.warnings + ' warning' : '')
-        + '</div></div>';
-    }
-
-    // Build findings table
-    var findingsHtml = "";
-    // Sort: critical first, then warning, info, pass
-    var sortedFindings = report.findings.slice().sort(function (a, b) {
-      var order = { critical: 0, warning: 1, info: 2, pass: 3 };
-      return (order[a.severity] || 4) - (order[b.severity] || 4);
-    });
-
-    for (var fi = 0; fi < sortedFindings.length; fi++) {
-      var f = sortedFindings[fi];
-      var sc = severityColors[f.severity] || severityColors.info;
-      var recHtml = f.recommendation
-        ? '<div style="margin-top:6px;padding:8px 12px;background:' + (dark ? "#1a1a2e" : "#f1f3f5")
-          + ';border-left:3px solid ' + sc.bg + ';border-radius:4px;font-size:13px">💡 ' + escapeHtml(f.recommendation) + '</div>'
-        : '';
-
-      findingsHtml += '<div style="padding:14px 16px;border-bottom:1px solid ' + borderColor + '">'
-        + '<div style="display:flex;align-items:center;gap:10px">'
-        + '<span style="display:inline-block;padding:2px 10px;border-radius:12px;font-size:11px;font-weight:700;text-transform:uppercase;background:'
-        + sc.bg + ';color:' + sc.text + '">' + f.severity + '</span>'
-        + '<span style="font-size:12px;color:' + mutedText + '">' + escapeHtml(f.id) + '</span>'
-        + '<span style="font-weight:600">' + escapeHtml(f.title) + '</span>'
-        + '</div>'
-        + '<div style="margin-top:4px;font-size:13px;color:' + mutedText + '">' + escapeHtml(f.description) + '</div>'
-        + recHtml
-        + '</div>';
-    }
-
-    // Summary counters
-    var summaryItems = [
-      { label: "Passed", value: report.passed, color: "#28a745" },
-      { label: "Critical", value: report.criticals, color: "#dc3545" },
-      { label: "Warnings", value: report.warnings, color: "#ffc107" },
-      { label: "Total Checks", value: report.totalFindings, color: mutedText }
-    ];
-    var summaryHtml = "";
-    for (var si = 0; si < summaryItems.length; si++) {
-      var s = summaryItems[si];
-      summaryHtml += '<div style="text-align:center;flex:1;min-width:100px">'
-        + '<div style="font-size:28px;font-weight:700;color:' + s.color + '">' + s.value + '</div>'
-        + '<div style="font-size:12px;color:' + mutedText + '">' + s.label + '</div></div>';
-    }
-
-    var html = '<!DOCTYPE html><html lang="en"><head><meta charset="utf-8">'
-      + '<meta name="viewport" content="width=device-width,initial-scale=1">'
-      + '<title>' + escapeHtml(title) + '</title>'
-      + '<style>'
-      + 'body{margin:0;padding:20px;font-family:-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,sans-serif;'
-      + 'background:' + bg + ';color:' + textColor + '}'
-      + '.card{background:' + cardBg + ';border-radius:12px;box-shadow:0 2px 8px rgba(0,0,0,' + (dark ? '0.3' : '0.08') + ');margin-bottom:20px;overflow:hidden}'
-      + '.card-header{padding:16px 20px;font-weight:600;font-size:16px;border-bottom:1px solid ' + borderColor + '}'
-      + '.card-body{padding:20px}'
-      + '@media print{body{background:#fff;color:#333}.card{box-shadow:none;border:1px solid #ddd}}'
-      + '</style></head><body>'
-      + '<div style="max-width:900px;margin:0 auto">'
-      // Header
-      + '<div style="text-align:center;margin-bottom:24px">'
-      + '<h1 style="margin:0 0 4px 0;font-size:24px">' + escapeHtml(report.system) + ' Compliance Report</h1>'
-      + (showTime ? '<div style="color:' + mutedText + ';font-size:13px">Generated ' + escapeHtml(report.generatedAt) + '</div>' : '')
-      + '</div>'
-      // Grade circle + summary
-      + '<div class="card"><div class="card-body" style="display:flex;align-items:center;gap:24px;flex-wrap:wrap">'
-      + '<div style="flex-shrink:0;text-align:center">'
-      + '<div style="width:90px;height:90px;border-radius:50%;border:4px solid ' + gradeColor
-      + ';display:flex;flex-direction:column;align-items:center;justify-content:center">'
-      + '<div style="font-size:32px;font-weight:800;color:' + gradeColor + '">' + report.grade + '</div>'
-      + '<div style="font-size:12px;color:' + mutedText + '">' + report.overallScore + '/100</div>'
-      + '</div></div>'
-      + '<div style="display:flex;flex-wrap:wrap;flex:1;gap:8px">' + summaryHtml + '</div>'
-      + '</div></div>'
-      // Category scores
-      + '<div class="card"><div class="card-header">Category Scores</div><div class="card-body">'
-      + scoreBarsHtml
-      + '</div></div>'
-      // Findings
-      + '<div class="card"><div class="card-header">Findings (' + report.totalFindings + ')</div>'
-      + findingsHtml
-      + '</div>'
-      // Footer
-      + '<div style="text-align:center;padding:16px;font-size:12px;color:' + mutedText + '">'
-      + 'Generated by gif-captcha compliance reporter'
-      + '</div></div></body></html>';
-
-    return html;
-  }
-
-  /**
-   * Escape HTML special characters to prevent XSS in generated reports.
-   * @param {string} str - Raw string
-   * @returns {string} HTML-safe string
-   */
-  function escapeHtml(str) {
-    if (typeof str !== "string") return "";
-    return str.replace(/&/g, "&amp;").replace(/</g, "&lt;").replace(/>/g, "&gt;")
-      .replace(/"/g, "&quot;").replace(/'/g, "&#39;");
-  }
-
-  return {
-    generateReport: generateReport,
-    getRecommendedConfig: getRecommendedConfig,
-    compareReports: compareReports,
-    formatReportText: formatReportText,
-    formatReportHtml: formatReportHtml,
-    SEVERITY: SEVERITY,
-    CATEGORY: CATEGORY
   };
 }
 
@@ -10460,618 +9047,6 @@ function createMetricsAggregator(options) {
     isAutoCapturing: isAutoCapturing,
     onAlert: onAlert,
     exportHistory: exportHistory,
-  };
-}
-
-
-/**
- * createTrustScoreEngine — Unified trust scoring that aggregates signals from
- * multiple modules (reputation, fingerprinting, bot detection, rate limiting,
- * response analysis, etc.) into a single 0.0–1.0 trust score per client.
- *
- * Analogous to reCAPTCHA v3's invisible score but transparent and configurable.
- *
- * @param {Object} [options]
- * @param {Object} [options.weights] - Signal name → weight (default: equal)
- * @param {Object} [options.thresholds] - Action thresholds
- * @param {number} [options.thresholds.block=0.2] - Below this → block
- * @param {number} [options.thresholds.challenge=0.5] - Below this → challenge
- * @param {number} [options.thresholds.pass=0.7] - Above this → pass (between challenge and pass → soft challenge)
- * @param {number} [options.cacheTtlMs=30000] - Cache TTL per client
- * @param {number} [options.maxClients=5000] - Max cached client entries
- * @param {number} [options.maxHistory=100] - Max score history per client
- * @param {number} [options.decayFactor=0.95] - Weight of historical average in blended score
- * @returns {Object} Trust score engine instance
- */
-function createTrustScoreEngine(options) {
-  options = options || {};
-
-  var defaultWeights = {
-    reputation: 1.0,
-    fingerprint: 1.0,
-    botDetection: 1.0,
-    rateLimit: 1.0,
-    responseQuality: 1.0,
-    behaviorEntropy: 1.0
-  };
-
-  var weights = {};
-  var userWeights = options.weights || {};
-  var wKeys = Object.keys(defaultWeights);
-  for (var wi = 0; wi < wKeys.length; wi++) {
-    var wk = wKeys[wi];
-    weights[wk] = typeof userWeights[wk] === "number" ? userWeights[wk] : defaultWeights[wk];
-  }
-  // Allow custom signal names not in defaults
-  var uKeys = Object.keys(userWeights);
-  for (var ui = 0; ui < uKeys.length; ui++) {
-    if (typeof weights[uKeys[ui]] === "undefined") {
-      weights[uKeys[ui]] = userWeights[uKeys[ui]];
-    }
-  }
-
-  var thresholds = options.thresholds || {};
-  var blockThreshold = typeof thresholds.block === "number" ? thresholds.block : 0.2;
-  var challengeThreshold = typeof thresholds.challenge === "number" ? thresholds.challenge : 0.5;
-  var passThreshold = typeof thresholds.pass === "number" ? thresholds.pass : 0.7;
-
-  var cacheTtlMs = typeof options.cacheTtlMs === "number" && options.cacheTtlMs >= 0 ? options.cacheTtlMs : 30000;
-  var maxClients = typeof options.maxClients === "number" && options.maxClients > 0 ? options.maxClients : 5000;
-  var maxHistory = typeof options.maxHistory === "number" && options.maxHistory > 0 ? options.maxHistory : 100;
-  var decayFactor = typeof options.decayFactor === "number" ? options.decayFactor : 0.7;
-  var anomalyDropThreshold = typeof options.anomalyDropThreshold === "number" ? options.anomalyDropThreshold : 0.3;
-
-  // clientId → { score, signals, action, timestamp, history: [{ score, timestamp }] }
-  var clients = Object.create(null);
-  var clientOrder = new LruTracker(); // O(1) LRU order
-  var totalEvaluations = 0;
-  var actionCounts = { block: 0, challenge: 0, softChallenge: 0, pass: 0 };
-
-  // Registered signal providers: name → function(clientId) → { score: 0-1, confidence: 0-1, detail: string }
-  var providers = Object.create(null);
-
-  function _evictIfNeeded() {
-    while (clientOrder.length > maxClients) {
-      var oldest = clientOrder.evictOldest();
-      if (oldest !== undefined) delete clients[oldest];
-    }
-  }
-
-  function _touchClient(clientId) {
-    clientOrder.push(clientId);
-  }
-
-  function _determineAction(score) {
-    if (score <= blockThreshold) return "block";
-    if (score <= challengeThreshold) return "challenge";
-    if (score <= passThreshold) return "softChallenge";
-    return "pass";
-  }
-
-  /**
-   * Register a signal provider function.
-   * @param {string} name - Signal name (must match a weight key to contribute)
-   * @param {Function} fn - function(clientId) → { score: 0-1, confidence?: 0-1, detail?: string }
-   */
-  function registerProvider(name, fn) {
-    if (typeof name !== "string" || !name) return;
-    if (typeof fn !== "function") return;
-    providers[name] = fn;
-  }
-
-  /**
-   * Remove a signal provider.
-   * @param {string} name
-   */
-  function unregisterProvider(name) {
-    delete providers[name];
-  }
-
-  /**
-   * Evaluate trust score for a client using all registered providers.
-   * @param {string} clientId
-   * @param {Object} [manualSignals] - Optional manual signal overrides: { signalName: 0-1 score }
-   * @returns {Object} { clientId, score, action, signals, breakdown, cached }
-   */
-  function evaluate(clientId, manualSignals) {
-    if (typeof clientId !== "string" || !clientId) {
-      return { clientId: "", score: 0, action: "block", signals: {}, breakdown: [], cached: false, error: "invalid clientId" };
-    }
-
-    // Check cache
-    var cached = clients[clientId];
-    if (cached && (_now() - cached.timestamp) < cacheTtlMs) {
-      _touchClient(clientId);
-      return {
-        clientId: clientId,
-        score: cached.score,
-        action: cached.action,
-        signals: cached.signals,
-        breakdown: cached.breakdown,
-        cached: true
-      };
-    }
-
-    manualSignals = manualSignals || {};
-    var signals = {};
-    var breakdown = [];
-
-    // Collect signals from providers
-    var pNames = Object.keys(providers);
-    for (var pi = 0; pi < pNames.length; pi++) {
-      var pName = pNames[pi];
-      try {
-        var result = providers[pName](clientId);
-        if (result && typeof result.score === "number") {
-          signals[pName] = {
-            score: _clamp(result.score, 0, 1),
-            confidence: typeof result.confidence === "number" ? _clamp(result.confidence, 0, 1) : 1,
-            detail: result.detail || null
-          };
-        }
-      } catch (e) {
-        // Provider error — skip this signal
-        signals[pName] = { score: 0.5, confidence: 0, detail: "provider error: " + (e.message || e), error: true };
-      }
-    }
-
-    // Apply manual signal overrides
-    var mKeys = Object.keys(manualSignals);
-    for (var mi = 0; mi < mKeys.length; mi++) {
-      var mName = mKeys[mi];
-      var mVal = manualSignals[mName];
-      if (typeof mVal === "number") {
-        signals[mName] = { score: _clamp(mVal, 0, 1), confidence: 1, detail: "manual override" };
-      } else if (mVal && typeof mVal.score === "number") {
-        signals[mName] = {
-          score: _clamp(mVal.score, 0, 1),
-          confidence: typeof mVal.confidence === "number" ? _clamp(mVal.confidence, 0, 1) : 1,
-          detail: mVal.detail || "manual override"
-        };
-      }
-    }
-
-    // Compute weighted score
-    var weightedSum = 0;
-    var totalWeight = 0;
-    var sigNames = Object.keys(signals);
-    for (var si = 0; si < sigNames.length; si++) {
-      var sName = sigNames[si];
-      var sig = signals[sName];
-      var w = typeof weights[sName] === "number" ? weights[sName] : 0;
-      if (w <= 0) continue;
-
-      // Scale weight by confidence
-      var effectiveWeight = w * sig.confidence;
-      weightedSum += sig.score * effectiveWeight;
-      totalWeight += effectiveWeight;
-
-      breakdown.push({
-        signal: sName,
-        score: sig.score,
-        weight: w,
-        confidence: sig.confidence,
-        effectiveWeight: effectiveWeight,
-        contribution: sig.score * effectiveWeight,
-        detail: sig.detail
-      });
-    }
-
-    var rawScore = totalWeight > 0 ? weightedSum / totalWeight : 0.5;
-
-    // Blend with historical average (recency-weighted)
-    var blendedScore = rawScore;
-    if (cached && cached.history && cached.history.length > 0) {
-      // Recency-weighted average: more recent scores count more.
-      // Weight_i = (i+1) / sum(1..n), so last entry gets highest weight.
-      var histLen = cached.history.length;
-      var weightDenom = (histLen * (histLen + 1)) / 2; // sum of 1..n
-      var histWeightedSum = 0;
-      for (var hi = 0; hi < histLen; hi++) {
-        histWeightedSum += cached.history[hi].score * (hi + 1);
-      }
-      var histAvg = histWeightedSum / weightDenom;
-
-      // Anomaly detection: if rawScore drops sharply from historical
-      // average, use a more aggressive weight for the current evaluation
-      // to react faster to behavioral changes (e.g., trust-washing attacks).
-      var effectiveDecay = decayFactor;
-      var drop = histAvg - rawScore;
-      if (drop > anomalyDropThreshold) {
-        // Scale decay down proportionally — bigger drop = less history weight
-        // At drop=0.3 with threshold=0.3: effectiveDecay = decayFactor * 0.5
-        // At drop=0.6: effectiveDecay = decayFactor * 0.25 (floor)
-        var dropRatio = Math.min(drop / anomalyDropThreshold, 2);
-        effectiveDecay = decayFactor * Math.max(0.25, 1 - dropRatio * 0.5);
-      }
-
-      blendedScore = rawScore * (1 - effectiveDecay) + histAvg * effectiveDecay;
-    }
-
-    var finalScore = _clamp(Math.round(blendedScore * 1000) / 1000, 0, 1);
-    var action = _determineAction(finalScore);
-
-    // Sort breakdown by contribution descending
-    breakdown.sort(function (a, b) { return b.contribution - a.contribution; });
-
-    // Update cache — store RAW scores in history (not blended) to prevent
-    // compounding feedback loop where blended-of-blended scores create
-    // exponential dampening of behavioral changes.
-    var history = (cached && cached.history) ? cached.history.slice() : [];
-    history.push({ score: rawScore, timestamp: _now() });
-    if (history.length > maxHistory) {
-      history = history.slice(history.length - maxHistory);
-    }
-
-    clients[clientId] = {
-      score: finalScore,
-      action: action,
-      signals: signals,
-      breakdown: breakdown,
-      timestamp: _now(),
-      history: history,
-      rawScore: rawScore
-    };
-    _touchClient(clientId);
-    _evictIfNeeded();
-
-    totalEvaluations++;
-    actionCounts[action] = (actionCounts[action] || 0) + 1;
-
-    return {
-      clientId: clientId,
-      score: finalScore,
-      rawScore: rawScore,
-      action: action,
-      signals: signals,
-      breakdown: breakdown,
-      cached: false
-    };
-  }
-
-  /**
-   * Evaluate multiple clients in batch.
-   * @param {string[]} clientIds
-   * @param {Object} [manualSignals] - Shared manual signals for all
-   * @returns {Object[]} Array of evaluation results
-   */
-  function batchEvaluate(clientIds, manualSignals) {
-    if (!Array.isArray(clientIds)) return [];
-    var results = [];
-    for (var i = 0; i < clientIds.length; i++) {
-      results.push(evaluate(clientIds[i], manualSignals));
-    }
-    return results;
-  }
-
-  /**
-   * Get the cached score for a client without re-evaluating.
-   * @param {string} clientId
-   * @returns {Object|null}
-   */
-  function getScore(clientId) {
-    var c = clients[clientId];
-    if (!c) return null;
-    return {
-      clientId: clientId,
-      score: c.score,
-      action: c.action,
-      age: _now() - c.timestamp,
-      stale: (_now() - c.timestamp) >= cacheTtlMs,
-      historyLength: c.history ? c.history.length : 0
-    };
-  }
-
-  /**
-   * Get score trend for a client.
-   * @param {string} clientId
-   * @param {number} [lastN] - Last N scores to analyze
-   * @returns {Object|null} { trend, slope, min, max, avg, scores }
-   */
-  function getScoreTrend(clientId, lastN) {
-    var c = clients[clientId];
-    if (!c || !c.history || c.history.length < 2) return null;
-
-    var hist = c.history;
-    if (typeof lastN === "number" && lastN > 0 && lastN < hist.length) {
-      hist = hist.slice(hist.length - lastN);
-    }
-
-    var scores = [];
-    var sum = 0;
-    var min = 1;
-    var max = 0;
-    for (var i = 0; i < hist.length; i++) {
-      var s = hist[i].score;
-      scores.push(s);
-      sum += s;
-      if (s < min) min = s;
-      if (s > max) max = s;
-    }
-    var avg = sum / scores.length;
-
-    // Simple linear regression for slope
-    var n = scores.length;
-    var sumX = 0, sumY = 0, sumXY = 0, sumXX = 0;
-    for (var j = 0; j < n; j++) {
-      sumX += j;
-      sumY += scores[j];
-      sumXY += j * scores[j];
-      sumXX += j * j;
-    }
-    var slope = n > 1 ? (n * sumXY - sumX * sumY) / (n * sumXX - sumX * sumX) : 0;
-
-    var trend = "stable";
-    if (slope > 0.01) trend = "improving";
-    else if (slope < -0.01) trend = "declining";
-
-    return {
-      trend: trend,
-      slope: Math.round(slope * 10000) / 10000,
-      min: min,
-      max: max,
-      avg: Math.round(avg * 1000) / 1000,
-      count: scores.length,
-      scores: scores
-    };
-  }
-
-  /**
-   * Invalidate cache for a client (force re-evaluation next time).
-   * @param {string} clientId
-   */
-  function invalidate(clientId) {
-    if (clients[clientId]) {
-      clients[clientId].timestamp = 0; // Mark stale
-    }
-  }
-
-  /**
-   * Clear all cached data for a client.
-   * @param {string} clientId
-   */
-  function clearClient(clientId) {
-    delete clients[clientId];
-    clientOrder.remove(clientId);
-  }
-
-  /**
-   * Update action thresholds dynamically.
-   * @param {Object} newThresholds - { block?, challenge?, pass? }
-   */
-  function setThresholds(newThresholds) {
-    if (!newThresholds) return;
-    if (typeof newThresholds.block === "number") blockThreshold = newThresholds.block;
-    if (typeof newThresholds.challenge === "number") challengeThreshold = newThresholds.challenge;
-    if (typeof newThresholds.pass === "number") passThreshold = newThresholds.pass;
-  }
-
-  /**
-   * Get current threshold configuration.
-   * @returns {Object}
-   */
-  function getThresholds() {
-    return { block: blockThreshold, challenge: challengeThreshold, pass: passThreshold };
-  }
-
-  /**
-   * Update signal weights dynamically.
-   * @param {Object} newWeights - { signalName: weight }
-   */
-  function setWeights(newWeights) {
-    if (!newWeights) return;
-    var k = Object.keys(newWeights);
-    for (var i = 0; i < k.length; i++) {
-      if (typeof newWeights[k[i]] === "number") {
-        weights[k[i]] = newWeights[k[i]];
-      }
-    }
-  }
-
-  /**
-   * Get current weight configuration.
-   * @returns {Object}
-   */
-  function getWeights() {
-    var copy = {};
-    var k = Object.keys(weights);
-    for (var i = 0; i < k.length; i++) copy[k[i]] = weights[k[i]];
-    return copy;
-  }
-
-  /**
-   * Get aggregate statistics.
-   * @returns {Object}
-   */
-  function getStats() {
-    var clientCount = clientOrder.length;
-    var scoreSum = 0;
-    var scoreCounts = { high: 0, medium: 0, low: 0, veryLow: 0 };
-    var order = clientOrder.toArray();
-
-    for (var i = 0; i < order.length; i++) {
-      var c = clients[order[i]];
-      if (c) {
-        scoreSum += c.score;
-        if (c.score > passThreshold) scoreCounts.high++;
-        else if (c.score > challengeThreshold) scoreCounts.medium++;
-        else if (c.score > blockThreshold) scoreCounts.low++;
-        else scoreCounts.veryLow++;
-      }
-    }
-
-    return {
-      totalEvaluations: totalEvaluations,
-      activeClients: clientCount,
-      averageScore: clientCount > 0 ? Math.round((scoreSum / clientCount) * 1000) / 1000 : 0,
-      actionCounts: {
-        block: actionCounts.block || 0,
-        challenge: actionCounts.challenge || 0,
-        softChallenge: actionCounts.softChallenge || 0,
-        pass: actionCounts.pass || 0
-      },
-      scoreBuckets: scoreCounts,
-      providerCount: Object.keys(providers).length,
-      weights: getWeights(),
-      thresholds: getThresholds()
-    };
-  }
-
-  /**
-   * Get clients below a score threshold (for monitoring).
-   * @param {number} [threshold=0.5]
-   * @returns {Object[]} Array of { clientId, score, action, age }
-   */
-  function getLowScoreClients(threshold) {
-    if (typeof threshold !== "number") threshold = 0.5;
-    var results = [];
-    var now = _now();
-    var order2 = clientOrder.toArray();
-    for (var i = 0; i < order2.length; i++) {
-      var id = order2[i];
-      var c = clients[id];
-      if (c && c.score < threshold) {
-        results.push({
-          clientId: id,
-          score: c.score,
-          action: c.action,
-          age: now - c.timestamp
-        });
-      }
-    }
-    results.sort(function (a, b) { return a.score - b.score; });
-    return results;
-  }
-
-  /**
-   * Compare two clients' trust profiles.
-   * @param {string} clientA
-   * @param {string} clientB
-   * @returns {Object|null}
-   */
-  function compareClients(clientA, clientB) {
-    var a = clients[clientA];
-    var b = clients[clientB];
-    if (!a || !b) return null;
-
-    var signalDiffs = [];
-    var allSignals = {};
-    var sk;
-    for (sk in a.signals) allSignals[sk] = true;
-    for (sk in b.signals) allSignals[sk] = true;
-    var sNames = Object.keys(allSignals);
-
-    for (var i = 0; i < sNames.length; i++) {
-      var sn = sNames[i];
-      var sa = a.signals[sn] ? a.signals[sn].score : null;
-      var sb = b.signals[sn] ? b.signals[sn].score : null;
-      signalDiffs.push({
-        signal: sn,
-        scoreA: sa,
-        scoreB: sb,
-        diff: sa !== null && sb !== null ? Math.round((sa - sb) * 1000) / 1000 : null
-      });
-    }
-
-    return {
-      clientA: { clientId: clientA, score: a.score, action: a.action },
-      clientB: { clientId: clientB, score: b.score, action: b.action },
-      scoreDiff: Math.round((a.score - b.score) * 1000) / 1000,
-      signalComparison: signalDiffs
-    };
-  }
-
-  /**
-   * Export engine state for persistence.
-   * @returns {Object}
-   */
-  function exportState() {
-    var state = {};
-    var order3 = clientOrder.toArray();
-    for (var i = 0; i < order3.length; i++) {
-      var id = order3[i];
-      var c = clients[id];
-      if (c) {
-        state[id] = {
-          score: c.score,
-          action: c.action,
-          history: c.history,
-          timestamp: c.timestamp
-        };
-      }
-    }
-    return {
-      clients: state,
-      stats: {
-        totalEvaluations: totalEvaluations,
-        actionCounts: actionCounts
-      }
-    };
-  }
-
-  /**
-   * Import engine state from persistence.
-   * @param {Object} state
-   */
-  function importState(state) {
-    if (!state || !state.clients) return;
-    var ids = Object.keys(state.clients);
-    for (var i = 0; i < ids.length; i++) {
-      var id = ids[i];
-      var s = state.clients[id];
-      clients[id] = {
-        score: s.score || 0,
-        action: s.action || "block",
-        signals: {},
-        breakdown: [],
-        timestamp: s.timestamp || 0,
-        history: s.history || [],
-        rawScore: s.score || 0
-      };
-      clientOrder.push(id);
-    }
-    _evictIfNeeded();
-    if (state.stats) {
-      totalEvaluations = state.stats.totalEvaluations || 0;
-      if (state.stats.actionCounts) {
-        var ac = state.stats.actionCounts;
-        actionCounts.block = ac.block || 0;
-        actionCounts.challenge = ac.challenge || 0;
-        actionCounts.softChallenge = ac.softChallenge || 0;
-        actionCounts.pass = ac.pass || 0;
-      }
-    }
-  }
-
-  /**
-   * Reset all state.
-   */
-  function reset() {
-    clients = Object.create(null);
-    clientOrder.clear();
-    totalEvaluations = 0;
-    actionCounts = { block: 0, challenge: 0, softChallenge: 0, pass: 0 };
-  }
-
-  return {
-    registerProvider: registerProvider,
-    unregisterProvider: unregisterProvider,
-    evaluate: evaluate,
-    batchEvaluate: batchEvaluate,
-    getScore: getScore,
-    getScoreTrend: getScoreTrend,
-    invalidate: invalidate,
-    clearClient: clearClient,
-    setThresholds: setThresholds,
-    getThresholds: getThresholds,
-    setWeights: setWeights,
-    getWeights: getWeights,
-    getStats: getStats,
-    getLowScoreClients: getLowScoreClients,
-    compareClients: compareClients,
-    exportState: exportState,
-    importState: importState,
-    reset: reset
   };
 }
 
@@ -13833,8 +11808,11 @@ function createDeviceCohortAnalyzer(options) {
 }
 
 
-// ── Honeypot Injector (loaded from separate file) ───────────────────
+// ── Extracted modules (loaded from separate files) ──────────────────
 var createHoneypotInjector = require("./honeypot-injector").createHoneypotInjector;
+var createComplianceReporter = require("./compliance-reporter").createComplianceReporter;
+var createABExperimentRunner = require("./ab-experiment-runner").createABExperimentRunner;
+var createTrustScoreEngine = require("./trust-score-engine").createTrustScoreEngine;
 var gifCaptcha = {
   sanitize: sanitize,
   createSanitizer: createSanitizer,
